+++ src/main/java/models/classdiagram/Aggregation.java
@@ -10,7 +10,7 @@ public class Aggregation extends Relationship {
     {
         super(type);
     }
-    public Aggregation(Component from, Component to, String label,int x,int y) {
+    public Aggregation(Component from, Component to, String label,double x,double y) {
         super(from, to, "Aggregation", label,x,y);
         System.out.println("Aggregation constructor ");
